修改了串口传输函数

+++ 移植 - 全项目/user/getdata.h
@@ -1,12 +1,14 @@
 #ifndef  __GETDATA_H_
 #define  __GETDATA_H_
 
+#include "stm32f4xx.h"                  // Device header
+
 // 四元数结构体
 typedef struct {
-    float w;  // 标量部分
-    float x;  // 向量部分 i
-    float y;  // 向量部分 j
-    float z;  // 向量部分 k
+    int16_t w;  // 标量部分
+    int16_t x;  // 向量部分 i
+    int16_t y;  // 向量部分 j
+    int16_t z;  // 向量部分 k
 } Quaternion;
 
 extern float q[4];
